Updated openstack/openstack Project: openstack-infra/shade  68a8d513dd0cd6c2f24642f2f02e4c61f3a863a7
Handle pagination for glance images

The default glance image list pagination seems to be about 20, which
means for v2 you really need to deal with pagination every time. It also
seems that the limit parameter does _not_ allow you to get more items
than the server default, so you can't just say "limit 100000" and be
done with it.

In order to accomplish this, we need to have the adapter stop trying to
return only the image list when there are other top level keys (so the
code can read the next link) and then do a loop requesting the next
link.

To make us even happier, glance returns the next link as '/v2/images' but
we have already set the adapter to 'https://example.com/v2' due to
version discovery. Since we're setting the endpoint_override on the
adapater, it treats that as the root, leaving us with
https://example.com/v2/v2/images. To deal with that, introduce a 'raw'
adapter which is bound to whatever is in the catalog, rather than
whatever we found through version discovery.

Change-Id: I030147e0275d0c4ee89588e21b5970f7d81800d3
Story: 2000837