Updated openstack/openstack Project: openstack/nova  95fd12b053bbf8a08283764a30bcc7586ded5134
Fix ksa mocking in test_cinderclient_unsupported_v1

This test was previously mocking the keystoneauth1
loading code globally which was getting randomly
trampled by other tests that were mocking the
same thing.

This changes the scope on the mock so it should be
isolated from other tests mocking out the same thing.

Change-Id: Id6c35c583edf0938b5d0178ab2cfb54277f0225f
Closes-Bug: #1654393