Updated openstack/openstack Project: openstack/nova  9c5c06e231652d85c74257fd3115ed256a425453
HTTP interface for resource providers by aggregates

In a new 1.3 microversion, the GET /resource_providers handler gains
support for a new query parameter 'member_of' which takes a value of
'in:' and a comma separated list of aggregate uuids, or a single
aggregate uuid.

The response is the list of resource providers that are associated
with any of those aggregates, or an empty list if there are none.

If in an old microversion, the query parameter is not accepted and a
400 is returned.

Change-Id: I82fc2003ce85dcadfecfea506e7d4adb47258c7a