Updated openstack/openstack Project: openstack/neutron  499f3c6d56072506ce5bbdafca6f00e25a102ab5
Reduce IP link show calls for SR-IOV scan loop

The device scanning logic was calling 'ip link show' for every
vf_index on the system, which was taking an excessive amount of
time an CPU cycles on very dense SR-IOV servers (>64 interfaces).
This was being executed every 2 seconds so it would cause the
agent to take a large percentage of the CPU time.

This patch adjusts the logic to use two 'ip link show' commands for
checking all of the operations in a single get_assigned_devices_info
call.

Closes-Bug: #1653880
Change-Id: I361caaf7f70580ee8c7b48a4a9c9477bdefd8b4b