Updated openstack/openstack Project: openstack-infra/tripleo-ci  c64c0ec60812f207852d1afb697b3c74fa6eaf85
Collect the deploy.env and annotate appended sections

Saving this will be useful for debugging, and to grab when you want to run
deploy.sh locally. We also put the caller as a comment in the file
as it's appended with every call to echo_vars_to_deploy_env so it's easier
to see where mutations to the environment happen, and what to remove
when running deploy.sh locally.

Change-Id: I7470c8a5a539c50d05342ac57e176156cb8ec0ff