Updated openstack/openstack Project: openstack/neutron  4ef443b19d70cca3ab99f74084422f9f4c1c990c
Raise AssertionError instead of eventlet.timeout.Timeout when failing.

The eventlet.timeout.Timeout exception subclasses Python's BaseException.
This is problematic because that is a higher category exception than
Python's Exception. As a result of this type of exception being raised
during testing, this halts the normal test flow.
AssertionError will properly mark a test as failed.

Change-Id: Iaf74f877f6a70c0383d539e9b01d6db2ec0ddeca
Closes-Bug: 1654283