Updated openstack/openstack Project: openstack/nova  b1ee1db0be43f1a7dcb0add9900a251c524566a0
Trivial indentation fix

This fixes one of those cases where "visual" indentation results in
legal, but very ugly Python, as it makes the continued line indented to
the same level as the block below it. Indentation has semantic meaning
in Python, so two adjacent lines in different blocks should never be
indented to the same level.

Change-Id: I6014980d3aebcac60979852f6d571146d7b47042