Updated openstack/openstack Project: openstack/python-ironicclient  a07cbc907481a735f4a7cc29440562c33db6c82e
Raise on NodeManager get when invalid identifier provided

When calling the NodeManager get method with resource_id
equal to [], {}, (), None, "" or False the client returns a bad
data structure.

In order to be consistent with the API and the CLI the
resource_id is now validated before calling the API
and raises ValidationError in case of invalid value.

Change-Id: Idb2b424ab17e79da948595fb74df3ef9d81732c0
Closes-Bug: #1652142