Updated openstack/openstack Project: openstack-infra/tripleo-ci  fe8a8e2bfc1baf422eea9087389eaa4b8df4ef39
Add "undecloud" or extra node's details to OVB's nodes' json

If we tell the create-env script to create an extra node (with OVB's
undercloud environment), this tells the build-nodes-json command that it
should add the details for this extra node as well. This is added in an
extra section that is ignored by ironic.

Depends on https://github.com/cybertron/openstack-virtual-baremetal/pull/33

Change-Id: I77589b4e9b97c9bae6d695213b1327dcc86fe8aa