Updated openstack/openstack Project: openstack-infra/project-config  a87fa0d0d7760d7826c07d1c69f150867f42c52c
Run fedora platform job against devstack-gate

We are making changes for things like python3 and need to be able to
check that fedora continues to work with these devstack-gate changes.
Add the fedora platform job to devstack gate as a nonvoting check job
(same way it is set up on devstack) to ensure that we don't break fedora
with any changes.

Change-Id: I2eaed33b1740d6aa39f6865193f213b3ed4d2228