Updated openstack/openstack Project: openstack-infra/devstack-gate  2fd21d07b571e8d9ea1f88bdb5282fcb9a86fc05
Add save_* functions for clearer saving of artifacts

The cleanup function has a log of if [[ -r ... ]] logic around files
to be able to operate in a non fatal way. This creates a couple of
functions that make saving a file or directory more straight forward.

It migrates a number of existing calls to this model.

Co-Authored-By: YAMAMOTO Takashi <yamamoto@midokura.com>
Change-Id: I9ad412fc8a4ba7b046f2fa95226ca0d5fc796b23