Updated openstack/openstack Project: openstack-infra/project-config  15bb9abe23e9e2d63e5cdafb982067347362522c
neutron: kill non-dvr OVS scenario

The project is moving into DVR+HA direction for all gate jobs. At this
point, adding jobs that are not DVR+HA is counterproductive.

Of course, since Linuxbridge does not work with DVR yet, we keep that
other scenario job flavour as-is. But there is no excuse to care about
that non-dvr OVS job.

Change-Id: I7b029acd19d8705882792a0aa6ae4aac68a4253e