Updated openstack/openstack Project: openstack/manila  75e685d40d4e02d38e5dfa1700b4fa6734291073
Properly deprecate service_instance_network_helper_type

Change [1] removed the ``service_instance_network_helper_type``
option as part of the removal of nova network helper support
for the service instance module.

Put the option back in but deprecate it so that operators with
this configuration can see that it is no longer used.

TrivialFix

[1] I846b760fa7c01f7f86768399a2bfad9ced7e57cd

Change-Id: I14768a237b9d7ebc5aa55733bc6bc0531ec3fbb4