Updated openstack/openstack Project: openstack/neutron  232f862d0e6df00899b7597274fed346de431f1b
Stop using .delete() in external net handling code

This can prevent revision bumping on a network update if the
only thing that happened was the removal of the external network
model.
This also stops the use of .delete() for the external network
RBAC entries because they are tracked by the quota engine and the
quota engine is based on SQLAlchemy event handlers.

Change-Id: Iafd1a07e693d5719437c10325739693a88e2a52f