Updated openstack/openstack Project: openstack-infra/tripleo-ci  18a3b8991d9a1890b0ddda1cb6efa3db944b8e8e
Install openstackclient on te-broker

Ib01a0b4ee76c033739ce9f7003070aa29714875e makes use of osc, but it
turns out it wasn't installed on the te-broker.  This change adds
it to the system setup package list.

Note that I've already installed it manually on the existing
te-broker instance.  This is just to ensure it doesn't get lost if
we have to rebuild it.

Change-Id: I34da1da71b7ab2ac5890ef4a443079113baace05