Updated openstack/openstack Project: openstack/neutron  0e186449ce0d1f019ec1962b44fce4110ad05cb5
Protect against '.delete()' for quota and revisions

The devref was suggesting to do exactly the wrong thing for cases where
we have SQLAlchemy event listeners waiting for deleted objects.
DELETE statements fail to trigger any kind of SQLAlchemy event
listeners and the majority of our delete operations are small
(<10 items at once) so the performance difference is negligible
(especially since we do it so infrequently).

Change-Id: Id142179418a7b94d8d9695871b3fcd5dcc64730c