Updated openstack/openstack Project: openstack/requirements  f0fb896f4416fe100b18826444b6583a42fab46e
Avoid Jinja2 versions 2.9.[0-4]

There is a bug within ansible and/or jinja2 that causes an exception
when a string contains curly braces. A bug has already been opened
with the Ansible team to review the bug.

This patch pins the jinja2 release at a version less than 2.9. The
issue first appeared in the 2.9 release and is still present in
2.9.4 (latest available).

Jinja2 bug: pallets/jinja#654
Ansible bug: ansible/ansible#20098
Related-Bug: 1655397
Change-Id: I82561d3d738147683499697883d9d03ceb897f3d