Updated openstack/openstack Project: openstack-infra/shade  114d4af1407500374ec6d525661c217e0b844807
Rework limits normalization

There isn't good docs or a guide on what we're looking for with
normalization - so rather than nitpick a review to death, I just made a
patch. We should make a doc ...

Changes included:
- We need to be explicit about every key we're going to support. Some of
  the limits keys are deprecated, so just don't return them.
- Change them all from camel case to underscore to match other
  resources.
- Document them in model.rst
- Change the name from limits to compute_limits, since it's only limits
  for the compute service.
- Allow usage by a normal user (it's not an admin function)

There will be a follow up patch to convert this to direct rest calls,
which will be fun due to the ability of this to work against other
projects.

Change-Id: Icfb118d4289263c0dd906f600e370242f191f708