Updated openstack/openstack Project: openstack/cinder  0c9e84799ac3f50e4a2996b62d489dfe145497d6
Set replication_status automatically on retype

When retyping a volume we should be updating the replication_status in
accordance with the new volume type, but right now we are leaving it to
the drivers to do so, which means that in some cases it is not being
updated.

This patch fixes this by doing it automatically based on the
replication_enabled value in extra_specs.

Closes-Bug: #1643884
Change-Id: I10a7931ed4a73dfd2b69f0db979bc32a71aedb11