Updated openstack/openstack Project: openstack-infra/puppet-zuul  ace0412fb7cc66d295b457b83a7f0f9c6f012689
Install ansible 2.1.4.0 RC1 on launchers

This installs ansible 2.1.4.0 RC1 which includes some useful bug fixes
on the zuul launchers. Note that we couldn't easily use the pip package
provider for this beacuse ansible does not publish ansible RCs to pypi.
So we replace the package resource with an exec that invokes pip
directly to do what we want.

Change-Id: If49a64540258b3a7a7be4df7e86e23b14e019c52