Updated openstack/openstack Project: openstack/cinder  f58aa5de0ab80af4cbf7d18be419ddea2c8aa4af
Fix create_group with group type name

When creating a group with a group type name, an exception
was raised: AttributeError: 'dict' object has no attribute 'id'
This is because group_types.get_group_type_by_name returns a
dictionary of a group type, not an object. This patch fixes it.

Change-Id: I36c0ccd6262dfb392241b2898f57ecaef5dc5162
Closes-Bug: #1654098