Updated openstack/openstack Project: openstack/ironic-specs  4ecb418578e30615f4a69cd0b843695b670c3bc8
Add spec for OpenBMC driver

Currently IPMI is not fully supported by OpenBMC. Instead,
OpenBMC's power and boot device is controlled through a
RESTful interface. While IPMI may be a supported mechanism
in the future for OpenBMC, the recommended way to interact
with it will continue to be its RESTful API. This means that
servers running OpenBMC will require a new module to control
its power, a new module to set the boot device, and a hardware
type that has these modules supported. This patch adds
a spec for these items.

Partial-bug: #1634635
Change-Id: Ie272f3ad2d4ae5bcb1bd542007dfcdb4adfcb799