Updated openstack/openstack Project: openstack/tripleo-heat-templates  d6d65a18653f8fd1ce073b2a3eec2cba1cb4ae8a
Update net-config-noop to use apply-config

Some deployments of the undercloud would like to use
net-config-noop (to disable all network configuration).
This patch changes this noop software deployment to use
'apply-config' instead of 'os-apply-config' which allows it
to work with the undercloud installer which only supports the
newer (apply-config) style due to heat signalling requirements.

Change-Id: If7b579b759477afd81ad64a4847169e59ddcbd77