Updated openstack/openstack Project: openstack-dev/grenade  24a2dccc71b68b47d75ede54b253b5061894d0bb
local.conf support

This attempts to provide local.conf support by making it possible to
have local.conf files during initial processing and concattenating
them to the end of the final local.conf.

This is not fool proof, but given the limited ways that we currently
use configs in devstack-gate, it's good enough to start the
transition.

Depends-On: Ia19d64ba240adc642284e4b942f6ce12310f862b

Change-Id: I3347c0886b1c2c8c56c902f485f45fb126446451