Updated openstack/openstack Project: openstack/neutron  1e2cfe8c06fcc2df52daa77b2b767ed6ffc2b19f
Use the session loader in keystoneauth1 for designate

Using the session loader has the benefit of compatibility with
settings in other sections (like keystone_authtoken), and the
ability to use client certs and setting the timeout. This changes
the designate.ca_cert setting to designate.cafile, but the former
is added as a deprecated option, so existing config files will work.

DocImpact
ca_cert in [designate] is deprecated, use cafile instead.

Change-Id: I9f2173b02af5c3929a96ef8c773d587e9b673d62