Updated openstack/openstack Project: openstack/nova  146d2976433d5aaa3a67211f3a2b6666cc6f5a7a
Merge v21 view builder for flavor API

Now v2 and v2.1 code is merged and we do not need
to maintain the different set of view builder.
Previously there were different way of building the
complete response for flavor API. v2 used to do with extension
and v2.1 with view builder itself.

Part of bp:api-no-more-extensions

Change-Id: I4f8b6b6270ffc5aa64b5c9b936f61e0e96f7fe5b