Updated openstack/openstack Project: openstack-infra/project-config  e1a4d7b57e396adab080418c3ce261fab96e903d
Remove sheepdog jobs from cinder queue

The sheepdog driver currently has an issue with its new license that
no longer qualifies it to have CI run by infra resources. The driver
is also going to be marked as unsupported in Cinder and removed in
the Pike release if we can't resolve the licensing concerns.

Change-Id: I49dab60f34658a352c90b2b5ac2a4d8ea3922e97