Updated openstack/openstack Project: openstack/oslo.messaging  8424dbc7cb3f102e4a0841684c4dc14267cd3da9
[zmq] Dynamic connections failover

ZeroMQ manages failover automatically when socket is
being connected to multiple hosts. Dynamic connections
mode assumes to connect once per message so we connect
only to a single host. However we cannot say for sure
if this host is still alive. For failover reasons
we can connect to a primary host and then add some
more hosts as failover and let ZeroMQ to decide where
it finally sends the message.

Change-Id: I19cbb75aaea8a0b725dd6a8ff665392738e164a1