Updated openstack/openstack Project: openstack/oslo.context  e2b5b8f5d7bdd7ff6c8bfa0815fa78fd48fa835b
Add the service token details to context

In the case of communications that include service tokens we need to add
that information to the context so that we can also enforce policy on
it. Add this information and load this information in the from_environ
method. Add these details to to_policy_values so that we can start to
enforce policy based on these attributes.

Change-Id: Id90f32795905112de804a18ddc8a69c038c829bb