Updated openstack/openstack Project: openstack/requirements  60f2e7c15720460aef243ecb05ae0e5e31a1afe8
Add zhmcclient to global-requirements

zhmcclient is a python binding for the IBM z Systems
ReST API in DPM (Dynamic Parition Manager) mode. zhmcclient is used
by nova-dpm [1] and networking-dpm [2].

Nova-dpm contains the dpm virt driver for nova compute. It uses
zhmcclient to CRUDL instances.

Networking-dpm contains a mechanism driver for the neutron server
and the dpm neutron l2 agent. It requires zhmcclient for networking
related configuration.

Neutron-l2 agent and nova-compute both run on the compute node. The
compute node can be any system that has access to the HMC ReST
API. This can be another z Systems, but also an x86 system.
The controller node (with the dpm neutron mechanism driver) does not
need access to zhmcclient.

- Is the library already packaged in the distros we target (Ubuntu
  latest / Fedora latest)?
  No!

- Is the library actively maintained?
  yes!
  Commits history: https://github.com/zhmcclient/python-zhmcclient/commits/master
  Release history: https://pypi.org/project/zhmcclient/#history
  # of releases in 2016: 9

- Is the library good code?
  Travis testing (unittest) for various python version in place:
  https://travis-ci.org/zhmcclient/python-zhmcclient
  Test Coverage: ~86% (https://coveralls.io/github/zhmcclient/python-zhmcclient)

- Is the library python 3 compatible?
  Yes! Python 3.4 and 3.5 test jobs in place

- Is the library license compatible?
  yes! Apache 2.0 license

- Is the function of this library already covered by other libraries
  in ``global-requirements.txt``?
  No!

- Is the library required for OpenStack project or related dev or
  infrastructure setup?
  yes!
  * networking-dpm [1]
  * nova-dpm [2]

[1] https://launchpad.net/networking-dpm
[2] https://launchpad.net/nova-dpm

Change-Id: Ied0654e01ef8a3c069dec7645e903ad4b421502a