Updated openstack/openstack Project: openstack/keystone-specs  97984a5a6c93d1f3ec244c48d1d09cf4fedfdc33
Update per-user-MFA spec to represent new db table not column

The new rules will not simply be a column on user but a new table
loaded via ORM relationship(s). This is done for future proofing
and allowing further metadata on the rules than can be allowed
without massive data migrations with a simple column addition.

Change-Id: I6ed238a3d470a5ce726c6399849ff68d7923cb28