Updated openstack/openstack Project: openstack/cinder  07d2dc0f7cbb686a10d73774734f546351145730
VMware: Set backend UUID to volume UUID

Currently we do not set the UUID of volume backing in vCenter
to volume UUID. We identify the volume in the backend by its
name. Querying the volume by name in vCenter is slow for large
deployments. This patch sets the UUID of the volume backing in
vCenter to Cinder volume UUID so that we can use vCenter
SearchIndex API to query the volume by UUID which is much
faster than query by name.

Partial-bug: #1600754
Change-Id: Ie14ae1c3d49f7fc9d3ee9dc8173b305dff57a105