Updated openstack/openstack Project: openstack-dev/devstack  b96c160d87095e4ea6637126b3dfb5558dc4c60a
lib/tempest: allow tweaking volume-feature-enabled/manage_snapshot

Only a few Cinder backends support the 'manage snapshot' feature. So
we need a feature flag here. Luckily the LVM driver does support this
feature so default the feature flag to True in devstack(/Gate) but
introduce a variable to tweak the config.

Change-Id: Ifcb9f91059f08bdf2faf2a8d65229aba5742ee1c
Depends-On: I77be1cf85a946bf72e852f6378f0d7b43af8023a