Updated openstack/openstack Project: openstack/nova  31774fa2285a29e7b711a90586d120aaab0b627d
Don't bypass cellsv1 replication if cellsv2 maps are in place

When we start creating the cellsv2 mappings in the cellsv1 devstack
job, we need to avoid that causing us to circumvent the top-level
merged replica of the cell databases. Doing this causes us to return
stale information that has not yet been updated by cells replication.

Change-Id: Icfbb17cce8ce8b03dc8b7b4ffb202db01e5218a6