Updated openstack/openstack Project: openstack/heat  d95295b37ff2ad90e7d36b0710f57fa623e7de31
Unit tests: Don't mock client_plugin for Zaqar subscription

Not only were we not using the mock object for anything, it had the effect
that the client_plugin's ignore_not_found would do nothing (all exceptions
would be suppressed when calling it; no exceptions would be suppressed when
using it as a context manager). So our delete tests were not really testing
what we thought.

Change-Id: I465aae26777ec7cd60a40d612b55a1731f7e6d9e